Move bit-twiddling helpers to their own file
This also deduplicates a repeated bit of code that iterates over bits
